requirements.txt
aadict @ file:///home/conda/feedstock_root/build_artifacts/aadict_1602526467604/work
absl-py==0.12.0
aiohttp @ file:///tmp/build/80754af9/aiohttp_1603204936805/work
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828496740/work
asset @ file:///home/conda/feedstock_root/build_artifacts/asset_1602618103350/work
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blis==0.7.4
boto==2.49.0
boto3 @ file:///tmp/build/80754af9/boto3_1603222673683/work
botocore @ file:///tmp/build/80754af9/botocore_1603297996270/work
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1600699165083/work
chardet==4.0.0
click==7.1.2
contextvars==2.4
cryptography @ file:///tmp/build/80754af9/cryptography_1601047419573/work
cymem==2.0.5
Cython @ file:///tmp/build/80754af9/cython_1594831565616/work
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gensim==3.8.0
globre @ file:///home/conda/feedstock_root/build_artifacts/globre_1602525575090/work
google-api-core @ file:///tmp/build/80754af9/google-api-core-split_1601307068854/work
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-cloud-core @ file:///tmp/build/80754af9/google-cloud-core_1602864831599/work
google-cloud-storage @ file:///tmp/build/80754af9/google-cloud-storage_1603300774238/work
google-crc32c @ file:///tmp/build/80754af9/google-crc32c_1601305113494/work
google-pasta==0.2.0
google-resumable-media @ file:///tmp/build/80754af9/google-resumable-media_1601995638063/work
googleapis-common-protos @ file:///tmp/build/80754af9/googleapis-common-protos-feedstock_1600455604686/work
grpcio==1.34.1
h5py==3.1.0
hearstPatterns==0.1.3
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.0.1
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206602906/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
jedi @ file:///tmp/build/80754af9/jedi_1598371618777/work
Jinja2==2.11.2
jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.0
mkl-service==2.3.0
multidict @ file:///tmp/build/80754af9/multidict_1600456401639/work
murmurhash==1.0.5
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914804165/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
notebook @ file:///tmp/build/80754af9/notebook_1601501581489/work
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pathy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1617386154532/work
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.17.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.7.4
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyhocon @ file:///home/conda/feedstock_root/build_artifacts/pyhocon_1605080802401/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141725711/work
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
regex @ file:///tmp/build/80754af9/regex_1602786665657/work
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.3
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1605985617292/work
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
Send2Trash==1.5.0
sip==4.19.24
six==1.15.0
smart-open @ file:///tmp/build/80754af9/smart_open_1602262876581/work
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.13.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==8.0.3
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
traitlets==4.3.3
typeguard==2.12.0
typer==0.3.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.4
wasabi==0.8.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl @ file:///tmp/build/80754af9/yarl_1598045023428/work
zipp==3.4.1
